+++ Mall/Mall/Classes/ThirdPart/BuriedPoint/UICollectionView+Hook.m
@@ -38,7 +38,7 @@
 
 - (void)hook_collectionView:(UICollectionView *)collectionView didSelectRowAtIndexPath:(NSIndexPath *)indexPath{
     // NSLog(@"%@ %@", self.class, [NSString stringWithFormat:@"%ld,%ld",indexPath.row,indexPath.section]);
-    [[HookObjcLog shareInstance] recordHookClass:self.class identifier:[NSString stringWithFormat:@"%ld,%ld",(long)indexPath.row,indexPath.section]];
+    [[HookObjcLog shareInstance] recordHookClass:self.class identifier:[NSString stringWithFormat:@"%ld,%ld",(long)indexPath.row,(long)indexPath.section]];
     [self hook_collectionView:collectionView didSelectRowAtIndexPath:indexPath];
 }
